Refactor controllers: inheritance, location
