--- app/controllers/users/posts_controller.rb
@@ -1,9 +1,0 @@
-module Users
-  class PostsController < UsersController
-    expose_decorated(:posts) { user.posts.ordered_by_desc.page(params[:page]) }
-
-    def index
-      render "users/posts/index", posts: posts
-    end
-  end
-end
